# pt-general-go/docs/swagger.yaml
basePath: /general
definitions:
  domain.Activity:
    properties:
      activity:
        type: string
      iconName:
        type: string
    type: object
...
  domain.DifficultyLevel:
    enum:
    - EASY
    - MEDIUM
    - HARD
    type: string
    x-enum-varnames:
    - DifficultyLevelEASY
    - DifficultyLevelMEDIUM
    - DifficultyLevelHARD
...
  domain.Photo:
    properties:
      createdAt:
        type: string
      description:
        type: string
      id:
        type: string
      tourId:
        type: string
      url:
        type: string
    type: object
...
  domain.Review:
    properties:
      comment:
        type: string
      createdAt:
        type: string
      id:
        type: string
      image:
        type: string
      link:
        type: string
      rating:
        type: integer
      tourId:
        type: string
      userId:
        type: string
      userName:
        type: string
    type: object
  domain.Tour:
    properties:
      availableMonths:
        items:
          type: string
        type: array
      coverUrl:
        type: string
      createdAt:
        type: string
      ctaDescription:
        type: string
      ctaTitle:
        type: string
      description:
        type: string
      difficulty:
        $ref: '#/definitions/domain.DifficultyLevel'
      durationDays:
        type: integer
      endLocation:
        type: string
      faq:
        type: object
      groupSize:
        type: integer
      id:
        type: string
      languages:
        items:
          type: string
        type: array
      location:
        type: string
      minAge:
        type: integer
      popUp1Description:
        type: string
      popUp1ImageUrl:
        type: string
      popUp1Title:
        type: string
      popUp2Description:
        type: string
      popUp2ImageUrl:
        type: string
      popUp2Title:
        type: string
      price:
        type: number
      program:
        type: object
      reviewsSectionName:
        type: string
      slug:
        type: string
      spotsLeft:
        type: integer
      startLocation:
        type: string
      subtitle:
        type: string
      title:
        type: string
      updatedAt:
        type: string
    type: object
  domain.TourDate:
    properties:
      createdAt:
        type: string
      dateFrom:
        type: string
      dateTo:
        type: string
      groupSize:
        type: integer
      id:
        type: string
      isAvailable:
        type: boolean
      tourId:
        type: string
      updatedAt:
        type: string
    type: object
  domain.TourFull:
    properties:
      activities:
        items:
          $ref: '#/definitions/domain.Activity'
        type: array
      availableMonths:
        items:
          type: string
        type: array
      coverUrl:
        type: string
      createdAt:
        type: string
      ctaDescription:
        type: string
      ctaTitle:
        type: string
      dates:
        items:
          $ref: '#/definitions/domain.TourDate'
        type: array
      description:
        type: string
      difficulty:
        $ref: '#/definitions/domain.DifficultyLevel'
      durationDays:
        type: integer
      endLocation:
        type: string
      faq:
        type: object
      groupSize:
        type: integer
      id:
        type: string
      included:
        items:
          type: string
        type: array
      languages:
        items:
          type: string
        type: array
      location:
        type: string
      minAge:
        type: integer
      photos:
        items:
          $ref: '#/definitions/domain.Photo'
        type: array
      popUp1Description:
        type: string
      popUp1ImageUrl:
        type: string
      popUp1Title:
        type: string
      popUp2Description:
        type: string
      popUp2ImageUrl:
        type: string
      popUp2Title:
        type: string
      price:
        type: number
      program:
        type: object
      reviewAmount:
        type: integer
      reviews:
        items:
          $ref: '#/definitions/domain.Review'
        type: array
      reviewsSectionName:
        type: string
      slug:
        type: string
      spotsLeft:
        type: integer
      starAmount:
        type: number
      startLocation:
        type: string
      subtitle:
        type: string
      summary:
        items:
          type: string
        type: array
      title:
        type: string
      updatedAt:
        type: string
    type: object
...
  dto.CreateTourRequest:
    properties:
      availableMonths:
        example:
        - June
        - July
        - August
        items:
          type: string
        type: array
      coverUrl:
        example: https://example.com/cover.jpg
        type: string
      description:
        example: An exciting mountain tour
        type: string
      difficulty:
        example: MEDIUM
        type: string
      durationDays:
        example: 7
        type: integer
      endLocation:
        example: Astana
        type: string
      languages:
        example:
        - English
        - Russian
        items:
          type: string
        type: array
      minAge:
        example: 18
        type: integer
      price:
        example: 1500
        type: number
      program:
        type: object
      reviewsSectionName:
        example: Why travelers love this
        type: string
      slug:
        example: mountain-adventure
        type: string
      startLocation:
        example: Almaty
        type: string
      title:
        example: Mountain Adventure
        type: string
    required:
    - description
    - difficulty
    - title
    type: object
...
  dto.UpdateTourRequest:
    properties:
      availableMonths:
        example:
        - May
        - June
        - July
        - August
        - September
        items:
          type: string
        type: array
      coverUrl:
        example: https://example.com/new-cover.jpg
        type: string
      description:
        example: An updated exciting mountain tour
        type: string
      difficulty:
        example: HARD
        type: string
      durationDays:
        example: 10
        type: integer
      endLocation:
        example: Osh
        type: string
      languages:
        example:
        - English
        - Russian
        - Kazakh
        items:
          type: string
        type: array
      minAge:
        example: 21
        type: integer
      price:
        example: 2000
        type: number
      program:
        type: object
      reviewsSectionName:
        example: Why travelers love this
        type: string
      slug:
        example: mountain-adventure-updated
        type: string
      startLocation:
        example: Bishkek
        type: string
      title:
        example: Mountain Adventure Updated
        type: string
    type: object
...
host: localhost:8000
info:
  contact: {}
  description: API for PhotoTours platform
  title: PhotoTours API
  version: "1.0"
paths:
...
  /tours:
    get:
      consumes:
      - application/json
      description: Get a paginated list of tours with optional filters (location,
        date range, group size, price range, season)
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page (max 100)
        in: query
        name: limit
        type: integer
      - description: Filter by location
        in: query
        name: location
        type: string
      - description: Filter by start date (YYYY-MM-DD)
        in: query
        name: dateFrom
        type: string
      - description: Filter by end date (YYYY-MM-DD)
        in: query
        name: dateTo
        type: string
      - description: Filter by minimum group size
        in: query
        name: groupSize
        type: integer
      - description: Filter by minimum price
        in: query
        name: priceMin
        type: number
      - description: Filter by maximum price
        in: query
        name: priceMax
        type: number
      - description: Filter by season (winter, spring, summer, autumn)
        in: query
        name: season
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.TourFull'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all tours
      tags:
      - tours
    post:
      consumes:
      - application/json
      description: Create a new tour (requires guide or admin role)
      parameters:
      - description: Tour data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateTourRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.Tour'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new tour (Guide/Admin only)
      tags:
      - tours
  /tours/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a tour by ID (requires guide or admin role)
      parameters:
      - description: Tour ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete tour by ID (Guide/Admin only)
      tags:
      - tours
    get:
      consumes:
      - application/json
      description: Get full tour information by ID including guide, dates, photos,
        videos, materials, tags, and categories
      parameters:
      - description: Tour ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.TourFull'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get tour by ID
      tags:
      - tours
    patch:
      consumes:
      - application/json
      description: Update tour information by ID (requires guide or admin role)
      parameters:
      - description: Tour ID (UUID)
        in: path
        name: id
        required: true
        type: string
      - description: Update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateTourRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.Tour'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update tour by ID (Guide/Admin only)
      tags:
      - tours
  /tours/{id}/similar:
    get:
      consumes:
      - application/json
      description: Get similar tours for a given tour
      parameters:
      - description: Tour ID (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/domain.Tour'
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get similar tours by tour ID
      tags:
      - tours
  /tours/slug/{slug}:
    get:
      consumes:
      - application/json
      description: Get full tour information by slug including guide, dates, photos,
        videos, materials, tags, and categories
      parameters:
      - description: Tour slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.TourFull'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get tour by slug
      tags:
      - tours
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"